# backend/routes/sessionsRoutes.ts
import express, {Express, NextFunction, Request, Response} from 'express';
import { handleErrors } from './handleErrors';
import { PrismaClient } from '@prisma/client';
import bcrypt from 'bcrypt';
import { logEvent } from '../logger';
import configureXmlJsonSupport from '../middleware/xmljsonsupport';

const { v4: uuidv4 } = require('uuid');
const verifier = require('@gradeup/email-verify');
const prisma = new PrismaClient();
const router: Express = express(); // Change the type to Express

// Add XML/JSON support middleware
configureXmlJsonSupport(router);


// Routes
router.post(
    '/',
    handleErrors(async (req: Request, res: Response) => {
        try {
            const user = await prisma.user.findUnique({
                where: {
                    email: req.body.email,
                },
            });
            console.log(user);
            if (!user) return res.status(404).send('User not found');

            if (user.password === null) {
                return res.status(401).send('Password not set');
            }

            const passwordMatches = await bcrypt.compare(
                req.body.password,
                user.password
            );
            if (!passwordMatches) return res.status(401).send('Invalid password');

            const session = await prisma.session.create({
                data: { userId: user.id, id: uuidv4() },
            });

            logEvent('Sign-in', { userId: user.id });

            res.status(201).send(session);
        } catch (error) {
            console.error(error);
            res.status(500).send('Internal server error');
        }
    })
);

// Delete session
router.delete(
    '/',
    handleErrors(async (req: Request, res: Response) => {
        try {
            const sessionId = req.headers.authorization;
            if (!sessionId) {
                return res.status(401).send({error: 'Session not found'});
            }
            console.log('Session ID to delete:', sessionId);
            const session = await prisma.session.findUnique({
                where: { id: sessionId.substring(7) },
            });
            if (!session) {
                return res.status(404).send({error: 'Session not found'});
            }
            await prisma.session.delete({
                where: { id: sessionId.substring(7) },
            });

            logEvent('Sign-out', { sessionId: session.id });

            res.status(204).send(session);
        } catch (error) {
            console.error(error);
            res.status(500).send({error: 'Session not found'});
        }
    })
);

export default router;