# backend/swagger.yaml
openapi: 3.0.2
info:
  title: BirthdayBook
  version: 1.0.0
paths:
...
  /sessions:
    post:
      tags:
        - Sessions
      summary: Create a session
      description: Creates a new session for a user.
      requestBody:
        description: Credentials for user authentication.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
          application/xml:
            schema:
              $ref: '#/components/schemas/Credentials'
        required: true
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
            application/xml:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      operationId: createSession
      security:
        - Bearer: []

    delete:
      tags:
        - Sessions
      summary: Delete a session
      description: Deletes a session for a user.
      operationId: deleteSession
      responses:
        '204':
          description: Successful response. Session deleted.
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Unauthorized:
                  value:
                    error: Session not found
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Unauthorized:
                  value:
                    error: Session not found
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                SessionNotFound:
                  value:
                    error: Session not found
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                SessionNotFound:
                  value:
                    error: Session not found
        '500':
          description: Internal Server Error
      security:
        - Bearer: [ ]
components:
  schemas:
...
    Credentials:
      title: Root Type for Credentials
      description: ''
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: asd6@asd.ee
        password: '11111111'
    Error:
      title: Root Type for Error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: Email is required
...
    Session:
      title: Session
      type: object
      properties:
        sessionId:
          type: string
        userId:
          type: integer
      required:
        - sessionId
        - userId
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
...